Basic kernel portability
This changeset makes parts of the kernel work better on a 32bit system,
but defers arch specific changes for a later commit.

+++ usr/sys/sys/fio.c
@@ -103,7 +103,7 @@ openi(ip, rw)
 register struct inode *ip;
 {
 	dev_t dev;
-	register unsigned int maj;
+	register u16 maj;
 
 	dev = (dev_t)ip->i_un.i_rdev;
 	maj = major(dev);
